hax-compatibility: Cast match scrutinees

+++ libcrux-ml-dsa/src/sample.rs
@@ -210,7 +210,7 @@ pub(crate) fn rejection_sample_less_than_eta<SIMDUnit: Operations, const ETA: us
     sampled: &mut usize,
     out: &mut [i32; 263],
 ) -> bool {
-    match ETA {
+    match ETA as u8 {
         2 => rejection_sample_less_than_eta_equals_2::<SIMDUnit>(randomness, sampled, out),
         4 => rejection_sample_less_than_eta_equals_4::<SIMDUnit>(randomness, sampled, out),
         _ => unreachable!(),
@@ -337,7 +337,7 @@ fn sample_mask_ring_element<
 >(
     seed: [u8; 66],
 ) -> PolynomialRingElement<SIMDUnit> {
-    match GAMMA1_EXPONENT {
+    match GAMMA1_EXPONENT as u8 {
         17 => {
             let mut out = [0u8; 576];
             Shake256::shake256::<576>(&seed, &mut out);
@@ -374,7 +374,7 @@ pub(crate) fn sample_mask_vector<
     let seed2 = update_seed(seed, domain_separator);
     let seed3 = update_seed(seed, domain_separator);
 
-    match GAMMA1_EXPONENT {
+    match GAMMA1_EXPONENT as u8 {
         17 => {
             let mut out0 = [0; 576];
             let mut out1 = [0; 576];
